+++ src/main/java/blackjack/domain/scoreboard/result/GameResult.java
@@ -1,4 +1,4 @@
-package blackjack.domain.scoreboard;
+package blackjack.domain.scoreboard.result;
 
 import blackjack.domain.card.Card;
