+++ src/test/java/blackjack/domain/scoreboard/ScoreBoardTest.java
@@ -1,8 +1,10 @@
 package blackjack.domain.scoreboard;
 
 import blackjack.domain.card.Card;
-import blackjack.domain.card.Suit;
-import blackjack.domain.card.Value;
+import blackjack.domain.card.painting.Suit;
+import blackjack.domain.card.painting.Value;
+import blackjack.domain.scoreboard.result.GameResult;
+import blackjack.domain.scoreboard.result.UserGameResult;
 import blackjack.domain.user.Dealer;
 import blackjack.domain.user.User;
 import org.junit.jupiter.api.DisplayName;
