+++ src/test/java/blackjack/domain/user/DealerTest.java
@@ -1,8 +1,8 @@
 package blackjack.domain.user;
 
 import blackjack.domain.card.Card;
-import blackjack.domain.card.Suit;
-import blackjack.domain.card.Value;
+import blackjack.domain.card.painting.Suit;
+import blackjack.domain.card.painting.Value;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
