+++ src/test/java/blackjack/domain/user/StatusTest.java
@@ -1,5 +1,6 @@
 package blackjack.domain.user;
 
+import blackjack.domain.user.status.Status;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
